feat: refactor favorite entities and services to use ManyToOne relationships and update user entity properties
