+++ src/models/favorite/favorite.module.ts
@@ -1,4 +1,8 @@
 import { forwardRef, Module } from '@nestjs/common';
+import { TypeOrmModule } from '@nestjs/typeorm';
+import { FavAlbum } from 'src/db/fav/album';
+import { FavArtist } from 'src/db/fav/artist';
+import { FavTrack } from 'src/db/fav/track';
 import { AlbumModule } from '../album/album.module';
 import { ArtistModule } from '../artist/artist.module';
 import { TrackModule } from '../track/track.module';
@@ -7,12 +11,12 @@ import { FavoriteService } from './favorite.service';
 
 @Module({
   imports: [
+    TypeOrmModule.forFeature([FavAlbum, FavArtist, FavTrack]),
     forwardRef(() => AlbumModule),
     forwardRef(() => ArtistModule),
     forwardRef(() => TrackModule),
   ],
   controllers: [FavoriteController],
   providers: [FavoriteService],
-  exports: [FavoriteService],
 })
 export class FavoriteModule {}
